+++ test/ConfigurationTraitTest.php
@@ -14,6 +14,7 @@ use InteropTest\Config\TestAsset\ConnectionContainerIdConfiguration;
 use InteropTest\Config\TestAsset\ConnectionDefaultOptionsConfiguration;
 use InteropTest\Config\TestAsset\ConnectionMandatoryConfiguration;
 use InteropTest\Config\TestAsset\ConnectionMandatoryContainerIdConfiguration;
+use InteropTest\Config\TestAsset\ConnectionMandatoryRecursiveContainerIdConfiguration;
 use PHPUnit_Framework_TestCase as TestCase;
 
 /**
@@ -226,6 +227,27 @@ class ConfigurationTraitTest extends TestCase
 
         $config = $this->getTestConfig();
         unset($config['doctrine']['connection']['orm_default']['params']);
+
+        $stub->options($config);
+    }
+
+    /**
+     * Tests if options() throws a runtime exception if a recursive mandatory option is missing
+     *
+     * @covers \Interop\Config\ConfigurationTrait::options
+     */
+    public function testOptionsThrowsRuntimeExceptionIfMandatoryOptionRecursiveIsMissing()
+    {
+        $stub = new ConnectionMandatoryRecursiveContainerIdConfiguration();
+
+        $this->setExpectedException(
+            'Interop\Config\Exception\MandatoryOptionNotFoundException',
+            'Mandatory option "dbname"'
+        );
+
+        $config = $this->getTestConfig();
+
+        unset($config['doctrine']['connection']['orm_default']['params']['dbname']);
 
         $stub->options($config);
     }